added django_extensions added .env to gitignore, as we won't need it anymore
