+++ .gitignore
@@ -1,3 +1,3 @@
 __pycache__
-env
+.env
 *.sqlite3
